Split CLI helpers to restore main CI budget
